Avoid constantly changing sort order of openapi.json output.
The json serialization done in creation of k9-sak.openapi.json varied the order of some properties on each run. This caused the pipeline to always commit the changed output, even though there was no actual change to the api.

This change introduces use of a jackson ObjectMapper instance with more deterministic sorting of properties in output of the generated openapi spec file.

Hope this is sufficient to avoid unwanted change churn in the pipeline.